kartel-whygo-system/requirements.txt
# Core dependencies (existing)
python-dateutil==2.8.2

# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File handling
python-multipart==0.0.6

# Development
pytest==7.4.3
